Toast text fixed and renamed create/edit list overlay file
