+++ packages/esm-patient-list-app/translations/en.json
@@ -10,14 +10,15 @@
   "createNewPatientList": "Create new patient list",
   "delete": "Delete",
   "deleted": "Deleted",
-  "deletedPatientList": "Deleted Patient List",
+  "deletedPatientList": "Deleted patient list",
   "editList": "Edit list",
   "editNameDescription": "Edit Name/ Description",
   "editPatientListHeader": "Edit patient list",
   "error": "Error",
   "errorAddPatientToList": "Patient not added to list",
-  "errorCreatedPatientListDescription": "Couldn't edit patient list",
+  "errorCreatePatientListDescription": "Couldn't create patient list",
   "errorDeleteList": "Couldn't delete patient list",
+  "errorUpdatePatientListDescription": "Couldn't update patient list",
   "gender": "Gender",
   "identifier": "Identifier",
   "items": "items",
@@ -60,9 +61,9 @@
   "sex": "Sex",
   "startDate": "Start Date",
   "successAddPatientToList": "Patient added to list",
-  "successCreatedPatientList": "Created Patient List",
+  "successCreatedPatientList": "Created patient list",
   "successCreatedPatientListDescription": "Successfully created patient list",
-  "successEditedPatientList": "Editted Patient List",
-  "successEditPatientListDescription": "Successfully editted patient list",
-  "successfullyAdded": "Successfully added"
+  "successfullyAdded": "Successfully added",
+  "successUpdatePatientList": "Updated patient list",
+  "successUpdatePatientListDescription": "Successfully updated patient list"
 }
